backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic

# RAG & VectorDB
langchain
langchain-community
langchain-cohere
langchain-google-genai
langchain-huggingface
ollama
chromadb
sentence-transformers

# Document Loaders
pypdf
python-docx
unstructured

# Audio
openai-whisper
ffmpeg-python
python-multipart

# RAG Enhancement
langchainhub
cohere
flashrank

# Hugging Face Hub Utilities
hf_xet

# PyTorch
torch
#torch==2.7.1+cu118
#torchaudio==2.7.1+cu118
#torchvision==0.22.1+cu118